fix: Ledger and Account abstraction

+++ HDSLedger/Communication/src/main/java/pt/ulisboa/tecnico/hdsledger/communication/LedgerRequestTransfer.java
@@ -4,11 +4,11 @@ public class LedgerRequestTransfer extends Message {
 
     private String destinationId;
 
-    private int amount;
+    private long amount;
 
     private String signature;
 
-    public LedgerRequestTransfer(Type type, String senderId, String destinationId, int amount, String signature) {
+    public LedgerRequestTransfer(Type type, String senderId, String destinationId, long amount, String signature) {
         super(senderId, type);
         this.destinationId = destinationId;
         this.amount = amount;
@@ -19,9 +19,9 @@ public class LedgerRequestTransfer extends Message {
 
     public void setDestinationId(String destinationId) { this.destinationId = destinationId; }
 
-    public int getAmount() { return amount; }
+    public long getAmount() { return amount; }
 
-    public void setAmount(int amount) { this.amount = amount; }
+    public void setAmount(long amount) { this.amount = amount; }
 
     public String getSignature() { return signature; }
